feat(document): Add more functionality to DocumentProcessor and improve code documentation
In StringProcessor.cs, an 'override ToString()' method was added for returning processed content.

The majority of changes were made in DocumentProcessor.cs. The StreamReader is replaced with a BinaryReader for document reading; Constructors were modified and added to accommodate a BinaryReader. Finalizer and few other methods are added to the class. The major change is the addition of method JumpTo() which allows the document position to be changed in processing with correct state settings.

These changes reflect a necessity of handling different types of document data in the language assembler project.

IDocumentProcessor has been adjusted accordingly to include the JumpTo method.

Minor cleanup was done in document extensions and specific method documentation / comments were improved in DocumentLineInfo.cs.

The aim behind these changes is to improve the usability, flexibility, and comprehension of the code.

+++ src/LangAssembler/Extensions/StringProcessorExtensions.cs
@@ -1,6 +1,5 @@
 ﻿using System.Text;
 using LangAssembler.Options;
-using LangAssembler.Processors;
 using LangAssembler.Processors.Base;
 
 namespace LangAssembler.Extensions;
@@ -37,7 +36,7 @@ public static class StringProcessorExtensions
     /// </summary>
     /// <param name="processor">The processor instance to act upon.</param>
     /// <param name="range">The range to retrieve</param>
-    /// <returns>Substring of <see cref="IStringProcessor.Content"/></returns>
+    /// <returns>Substring of processor contents.</returns>
     public static string GetRange(this IStringProcessor processor, Range range) => processor[range];
     
     /// <summary>
@@ -196,7 +195,8 @@ public static class StringProcessorExtensions
     /// </summary>
     /// <param name="processor">The string processor</param>
     /// <returns>The character that was erased, or null if no character was removed.</returns>
-    public static int? ErasePrevious(this IEditableStringProcessor processor) => EraseChar(processor, (int) processor.Position - 1);
+    public static int? ErasePrevious(this IEditableStringProcessor processor) => 
+        EraseChar(processor, processor.Position - 1);
 
     /// <summary>
     /// Erases a specific character from the string processor.
